final homepage layout changes
